requirements.txt
networkx
shapely
psutil
seaborn
geopandas
gurobipy
websocket-client
kafka-python